refactor(client): dedupe location buffer + sketch handler logic
Extract the report location buffer-slider state into a shared
useLocationBuffer hook + BufferSlider component (consumed by the create
and confirm panels), and the map sketch/draw handlers into a shared
useSketchHandlers hook (consumed by the view and edit containers).
Behavior unchanged; removes the copy-pasted blocks Sonar flagged as new
duplicated lines on the PR.
