+++ client/src/containers/map/edit.tsx
@@ -1,26 +1,17 @@
 "use client";
 
-import { useCallback, useEffect, useMemo, useRef } from "react";
+import { useMemo } from "react";
 
 import dynamic from "next/dynamic";
 
-import { useAtom, useSetAtom } from "jotai";
+import { useAtomValue } from "jotai";
 
-import { getGeometryWithBuffer, useLocationGeometry } from "@/lib/location";
+import { useLocationGeometry } from "@/lib/location";
 
-import {
-  gridHoverAtom,
-  sketchActionAtom,
-  sketchAtom,
-  tmpBboxAtom,
-  useFormLocation,
-  useSyncGridSelectedDataset,
-  useSyncLocation,
-} from "@/app/(frontend)/store";
-
-import { BUFFERS } from "@/constants/map";
+import { tmpBboxAtom, useFormLocation, useSyncGridSelectedDataset } from "@/app/(frontend)/store";
 
 import GridLegend from "@/containers/map/grid-legend/grid";
+import { useSketchHandlers } from "@/containers/map/use-sketch-handlers";
 
 import Controls from "@/components/map/controls";
 import BasemapControl from "@/components/map/controls/basemap";
@@ -48,18 +39,24 @@ export default function MapEditContainer({
   desktop?: boolean;
   gridEnabled?: boolean;
 }) {
-  const [tmpBbox, setTmpBbox] = useAtom(tmpBboxAtom);
+  const tmpBbox = useAtomValue(tmpBboxAtom);
 
   const { location: defaultLocation } = useFormLocation();
   const GEOMETRY = useLocationGeometry(defaultLocation);
 
-  const [sketch, setSketch] = useAtom(sketchAtom);
-  const [sketchAction, setSketchAction] = useAtom(sketchActionAtom);
-  const sketchActionTimeoutRef = useRef<NodeJS.Timeout | null>(null);
-
-  const setGridHover = useSetAtom(gridHoverAtom);
-  const [location, setLocation] = useSyncLocation();
   const [gridSelectedDataset] = useSyncGridSelectedDataset();
+
+  const {
+    sketch,
+    sketchAction,
+    location,
+    handleCreate,
+    handleCreateChange,
+    handleCancel,
+    handleUpdate,
+    handleUpdateChange,
+    handlePointerLeave,
+  } = useSketchHandlers();
 
   const defaultBbox = useMemo(() => {
     if (GEOMETRY?.extent)
@@ -75,108 +72,6 @@ export default function MapEditContainer({
 
     return [-8999366.738755312, -4376503.729887867, -4792272.701940329, 2354846.7290161047];
   }, [GEOMETRY, desktop]);
-
-  const handleCreate = useCallback(
-    async (graphic: __esri.Graphic) => {
-      setSketch({ enabled: undefined, type: undefined });
-
-      if (graphic.geometry) {
-        setLocation({
-          type: graphic.geometry.type,
-          geometry: graphic.geometry.toJSON(),
-          buffer: BUFFERS[graphic.geometry.type],
-        });
-
-        const g = await getGeometryWithBuffer(graphic.geometry, BUFFERS[graphic.geometry.type]);
-        if (g?.extent) {
-          setTmpBbox(g.extent);
-        }
-      }
-
-      sketchActionTimeoutRef.current = setTimeout(() => {
-        setSketchAction({ type: undefined, state: undefined, geometryType: undefined });
-      }, 5000);
-    },
-    [setTmpBbox, setSketch, setLocation, setSketchAction],
-  );
-
-  const handleCreateChange = useCallback(
-    (e: __esri.SketchViewModelCreateEvent) => {
-      if (sketchActionTimeoutRef.current) {
-        clearTimeout(sketchActionTimeoutRef.current);
-      }
-      setSketchAction({ type: "create", state: e.state, geometryType: sketch.type });
-    },
-    [setSketchAction, sketch.type],
-  );
-
-  const handleCancel = useCallback(() => {
-    setSketch({ enabled: undefined, type: undefined });
-    setSketchAction({ type: undefined, state: undefined, geometryType: undefined });
-  }, [setSketch, setSketchAction]);
-
-  const handleUpdate = useCallback(
-    async (graphic: __esri.Graphic) => {
-      if (!location || !graphic.geometry) return;
-      const b = location.type !== "search" ? location.buffer : BUFFERS[graphic.geometry.type];
-      // Update the location state with the updated geometry
-      setLocation({
-        type: graphic.geometry.type,
-        geometry: graphic.geometry.toJSON(),
-        buffer: b,
-      });
-
-      // Optionally update the bounding box based on the updated geometry
-      const g = await getGeometryWithBuffer(graphic.geometry, b);
-      if (g?.extent) {
-        setTmpBbox(g.extent);
-      }
-
-      setSketch({ enabled: undefined, type: undefined });
-      setSketchAction({ type: undefined, state: undefined, geometryType: undefined });
-    },
-    [location, setLocation, setTmpBbox, setSketch, setSketchAction],
-  );
-
-  const handleUpdateChange = useCallback(
-    (e: __esri.SketchViewModelUpdateEvent) => {
-      if (sketchActionTimeoutRef.current) {
-        clearTimeout(sketchActionTimeoutRef.current);
-      }
-
-      if (!!e.graphics.length && e.graphics[0].geometry) {
-        setSketchAction({
-          type: "update",
-          state: e.state,
-          geometryType: e.graphics[0].geometry.type,
-        });
-      }
-    },
-    [setSketchAction],
-  );
-
-  const handlePointerLeave = useCallback(() => {
-    setGridHover({
-      id: null,
-      cell: undefined,
-      index: undefined,
-      values: [],
-      x: null,
-      y: null,
-      coordinates: undefined,
-    });
-  }, [setGridHover]);
-
-  useEffect(() => {
-    if (sketch.enabled === "edit") {
-      // focus map
-      const mapElement = document.querySelector("#map-default .esri-view-surface") as HTMLElement;
-
-      if (mapElement) {
-        mapElement.focus();
-      }
-    }
-  }, [sketch.enabled]);
 
   return (
     <div className="relative flex w-full grow flex-col">
